[ISSUE #9559]♻️Refactor lifecycle and transport modules; add compatibility module for NameServerUpdateCallback
