+++ rocketmq-client/src/implementation/mq_client_api_impl/transport/nameserver_callback.rs
@@ -1,0 +1,21 @@
+// Copyright 2023 The RocketMQ Rust Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use super::*;
+
+impl NameServerUpdateCallback for MQClientAPIImpl {
+    fn on_name_server_address_changed(&self, namesrv_address: Option<String>) -> String {
+        namesrv_address.unwrap_or_default()
+    }
+}
